MAJOR: refactor JSONDocument
introduced: JSONDocumentBase and JSONDocumentView

https://github.com/metasfresh/metasfresh-webui-api/issues/269
