+++ src/main/java/de/metas/ui/web/handlingunits/HUDocumentViewSelectionFactory.java
@@ -12,11 +12,11 @@ import de.metas.ui.web.view.DocumentViewFactory;
 import de.metas.ui.web.view.IDocumentViewSelection;
 import de.metas.ui.web.view.IDocumentViewSelectionFactory;
 import de.metas.ui.web.view.descriptor.DocumentViewLayout;
-import de.metas.ui.web.view.json.JSONCreateDocumentViewRequest;
+import de.metas.ui.web.view.json.JSONDocumentViewCreateRequest;
 import de.metas.ui.web.view.json.JSONDocumentViewLayout;
+import de.metas.ui.web.view.json.JSONViewDataType;
 import de.metas.ui.web.window.datatypes.DocumentPath;
 import de.metas.ui.web.window.datatypes.json.JSONOptions;
-import de.metas.ui.web.window.datatypes.json.JSONViewDataType;
 import de.metas.ui.web.window.descriptor.DocumentFieldWidgetType;
 import de.metas.ui.web.window.descriptor.DocumentLayoutElementDescriptor;
 import de.metas.ui.web.window.descriptor.DocumentLayoutElementFieldDescriptor;
@@ -120,7 +120,7 @@ public class HUDocumentViewSelectionFactory implements IDocumentViewSelectionFac
 	}
 
 	@Override
-	public IDocumentViewSelection createView(final JSONCreateDocumentViewRequest jsonRequest)
+	public IDocumentViewSelection createView(final JSONDocumentViewCreateRequest jsonRequest)
 	{
 		final String viewId = UUID.randomUUID().toString();
 		final int adWindowId = jsonRequest.getAD_Window_ID();
