+++ src/main/java/de/metas/ui/web/view/IDocumentViewSelectionFactory.java
@@ -1,9 +1,9 @@
 package de.metas.ui.web.view;
 
-import de.metas.ui.web.view.json.JSONCreateDocumentViewRequest;
+import de.metas.ui.web.view.json.JSONDocumentViewCreateRequest;
 import de.metas.ui.web.view.json.JSONDocumentViewLayout;
+import de.metas.ui.web.view.json.JSONViewDataType;
 import de.metas.ui.web.window.datatypes.json.JSONOptions;
-import de.metas.ui.web.window.datatypes.json.JSONViewDataType;
 
 /*
  * #%L
@@ -31,5 +31,5 @@ public interface IDocumentViewSelectionFactory
 {
 	JSONDocumentViewLayout getViewLayout(int adWindowId, JSONViewDataType viewDataType, JSONOptions jsonOpts);
 
-	IDocumentViewSelection createView(JSONCreateDocumentViewRequest jsonRequest);
+	IDocumentViewSelection createView(JSONDocumentViewCreateRequest jsonRequest);
 }
